moved all book models to 'book' package
